% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u5mr.R
\name{u5mr}
\alias{u5mr}
\title{Under five mortality rate (u5MR) estimation}
\usage{
u5mr(
  data,
  women = "women",
  child_born = "child_born",
  child_dead = "child_dead",
  agegrp = "agegrp",
  model = "CD West",
  svy_year,
  sex,
  e_0 = 60,
  mac = 28
)
}
\arguments{
\item{data}{A data.frame containing the information.}

\item{women}{A numeric variable with the women (N) number.}

\item{child_born}{A numeric variable with the child ever born (CEB) number.}

\item{child_dead}{A numeric variable with the child deaths (CD).}

\item{agegrp}{A character string of the mother age groups.}

\item{model}{A character string of the model life table.}

\item{svy_year}{A numeric variable with the reference date in decimals.}

\item{sex}{A character string with the sex.}

\item{e_0}{A numeric variable with the life expectancy corresponding to the life table selected.}

\item{mac}{A numeric variable with the mother average of child-bearing.}
}
\value{
A data frame with the q_1, q_5 and reference time estimated.
}
\description{
Estimate de u5MR from \code{data} values with the different versions of Brass method.
}
